fixes/improvements to lib code from work
- topology.cpp: modify interface due to thread-safety issue. caller is now responsible for ensuring the first _Detect call isn't reentered; everything else is safe.
- fix thread-safety issue in wnuma; use winit mechanism to ensure it's ready before use
- VFS now takes a cacheSize parameter (required for being able to disable read-only file caches for the image loader at work)
- allow dynarray that isn't actually holding memory 
- debug_stl: VC9 fix (disable this code except on the exact STL version on which it was tested)
- zlib, lib_api: changes to macro used to toggle between static and dynamic linking
- add boost filesystem header in external_libraries
- amd64: cpu_ topology functions are now provided by x86_x64
- cpu: remove cpu_ClockFrequency (dangerous, may be tempting to use during WHRT init which would cause a crash). use x86_x64_ClockFrequency or os_cpu_ClockFrequency instead.
- werrno: cope with newer boost versions
- wmman: follow SUSv3 in rejecting zero-length mappings

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5954 3db68df2-c116-0410-a063-a993310a9797

+++ source/network/NetLog.cpp
@@ -141,7 +141,7 @@ void CNetLogSink::DoSink( const CNetLogEvent& event )
 //-----------------------------------------------------------------------------
 void CNetLogSink::DoBulkSink( const CNetLogEvent* pEvents, size_t eventCount )
 {
-	unsigned*	pIndices	= NULL;
+	size_t*	pIndices	= NULL;
 	size_t	indexCount  = 0;
 	size_t	i;
